Extract shell.console from shell.client
The remaining two interfaces have been moved into shell.core
