+++ pom.xml
@@ -34,7 +34,7 @@
     </dependencyManagement>
 
     <modules>
-        <module>shell.client</module>
+        <module>shell.console</module>
         <module>shell.core</module>
     </modules>
