+++ shell.console/pom.xml
@@ -4,9 +4,9 @@
          xsi:schemaLocation="http://maven.apache.org/POM/4.0.0 http://maven.apache.org/xsd/maven-4.0.0.xsd">
     <modelVersion>4.0.0</modelVersion>
 
-    <artifactId>shell.client</artifactId>
-    <name>Client</name>
-    <description>Spoofax Shell client</description>
+    <artifactId>shell.console</artifactId>
+    <name>Console Client</name>
+    <description>Spoofax Shell console client</description>
     <packaging>jar</packaging>
 
     <parent>
@@ -56,7 +56,7 @@
                 <configuration>
                     <archive>
                         <manifest>
-                            <mainClass>org.metaborg.spoofax.shell.client.Repl</mainClass>
+                            <mainClass>org.metaborg.spoofax.shell.client.console.Repl</mainClass>
                         </manifest>
                     </archive>
                     <descriptorRefs>
