+++ shell.console/src/main/java/org/metaborg/spoofax/shell/client/console/OnEvalSuccessHook.java
@@ -1,26 +1,27 @@
-package org.metaborg.spoofax.shell.client;
+package org.metaborg.spoofax.shell.client.console;
 
 import java.util.function.Consumer;
+
+import org.metaborg.spoofax.shell.client.IDisplay;
 
 import com.google.inject.Inject;
 
 /**
  * Called upon success of an evaluation command.
  */
-public class OnEvalErrorHook implements Consumer<String> {
-
+class OnEvalSuccessHook implements Consumer<String> {
     private IDisplay display;
 
     /**
      * @param display The {@link IDisplay} to show the result on.
      */
     @Inject
-    OnEvalErrorHook(IDisplay display) {
+    OnEvalSuccessHook(IDisplay display) {
         this.display = display;
     }
 
     @Override
     public void accept(String s) {
-        display.displayError(s);
+        display.displayResult(s);
     }
 }
