+++ shell.core/src/main/java/org/metaborg/spoofax/shell/client/IDisplay.java
